.ajuste

+++ app/Http/Controllers/MarcaController.php
@@ -7,6 +7,11 @@ use Illuminate\Http\Request;
 
 class MarcaController extends Controller
 {
+
+    public function __construct(Marca $marca) {
+        $this->marca = $marca;
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -14,7 +19,7 @@ class MarcaController extends Controller
      */
     public function index()
     {
-        $marca = Marca::all();
+        $marca = $this->marca->all();
         return $marca;
     }
 
@@ -36,18 +41,22 @@ class MarcaController extends Controller
      */
     public function store(Request $request)
     {
-        $marca =  Marca::create($request->all());
+        $marca =  $this->marca->create($request->all());
         return $marca;
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  \App\Models\Marca  $marca
+     * @param  Integer
      * @return \Illuminate\Http\Response
      */
-    public function show(Marca $marca)
+    public function show($id)
     {
+        $marca = $this->marca->find($id);
+        if($marca === null){
+            return ['erro' => 'Recurso pesquisado não existe'];
+        }
         return $marca;
     }
 
@@ -66,11 +75,15 @@ class MarcaController extends Controller
      * Update the specified resource in storage.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Marca  $marca
+     * @param  Integer
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, Marca $marca)
+    public function update(Request $request, $id)
     {
+        $marca = $this->marca->find($id);
+        if($marca === null){
+            return ['erro' => 'Impossivel realizar a atualização, indice inexistente'];
+        }
         $marca->update($request->all());
         return $marca;
     }
@@ -78,11 +91,15 @@ class MarcaController extends Controller
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\Marca  $marca
+     * @param  Integer
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Marca $marca)
+    public function destroy($id)
     {
+        $marca = $this->marca->find($id);
+        if($marca === null){
+            return ['erro' => 'Impossivel realizar a exclusão. O recurso solicitado não existe'];
+        }
         $marca->delete();
         return ['msg' => 'A marca foi removida com sucesso'];
     }
